Move code from __init__ module to _extruct
The namespace would be cleaner if we keep only the imports in ``__init__``.
Not naming the new module ``extruct.extruct`` as this would interfere with
``extruct.extract`` autocompletion.
